Fix Clippy Dependencies
